refactor: WriteTabContent 컴포넌트 분리
