+++ pages/main/index.tsx
@@ -1,50 +1,19 @@
-import { useEffect, useRef, useState } from 'react';
+import { useState } from 'react';
 
-import WriteInput from '@components/Input/WriteInput';
 import Layout from '@components/Layout';
 import MainPageTab from '@components/Layout/MainPageTab';
-import WriteGuide from '@domain/끄적이는/components/Guide';
-import TemporalMemoHistoryTable from '@domain/끄적이는/components/History';
-import TodayTemoralMemos from '@domain/끄적이는/components/Today';
-import useCreateTemporalMemo from '@domain/끄적이는/mutations/useCreateTemporalMemo';
-import useGetWriteHistory from '@domain/끄적이는/queries/useGetHistory';
-import * as styles from '@domain/끄적이는/style.css';
+import WriteTabContent from '@domain/끄적이는/components/WriteTabContent';
 import 참고하는TabContent from '@domain/참고하는/components';
-import { useInput } from '@hooks/useInput';
-import useGetMemoFolders from '@queries/useGetMemoFolders';
 import useGetMyProfile from '@queries/useGetMyProfile';
 import { COLORS } from '@styles/tokens';
 
 const MainPage = () => {
   useGetMyProfile();
 
-  const writeContentRef = useRef<HTMLDivElement>(null);
-  const writeInput = useInput({ id: 'write-input' });
-  const { todayMemos, history } = useGetWriteHistory();
-  const { mutate: submitTemporalMemo } = useCreateTemporalMemo();
-  const { data: memoFolders } = useGetMemoFolders();
-
   const [selectedTab, setSelectedTab] = useState('끄적이는');
-
-  useEffect(() => {
-    handleScroll();
-  }, [todayMemos]);
-
-  const handleScroll = () => {
-    if (writeContentRef.current) {
-      writeContentRef.current.scrollTop = writeContentRef.current.scrollHeight;
-    }
-  };
 
   const handleTabSelect = (selectedTab: string) => {
     setSelectedTab(selectedTab);
-  };
-
-  const handleSubmit = () => {
-    submitTemporalMemo(writeInput.value);
-    writeInput.reset();
-
-    handleScroll();
   };
 
   const backgroundColor =
@@ -53,24 +22,7 @@ const MainPage = () => {
   return (
     <Layout backgroundColor={backgroundColor}>
       <MainPageTab
-        write={
-          <div className={styles.container}>
-            <div ref={writeContentRef} className={styles.content}>
-              <WriteGuide />
-              <TemporalMemoHistoryTable
-                data={history}
-                memoFolders={memoFolders}
-              />
-              <TodayTemoralMemos
-                memos={todayMemos[0]?.temporalMemos}
-                memoFolders={memoFolders}
-              />
-              <div className={styles.inputWrapper}>
-                <WriteInput inputProps={writeInput} onSubmit={handleSubmit} />
-              </div>
-            </div>
-          </div>
-        }
+        write={<WriteTabContent />}
         refer={<참고하는TabContent />}
         selectedTab={selectedTab}
         handleTabSelect={handleTabSelect}
