+++ src/domain/끄적이는/components/WriteTabContent/style.css.ts
